requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.15
django-bootstrap4==0.0.7
django-filter==1.1.0
django-jquery==3.1.0
django-toolbelt==0.0.1
djangorestframework==3.5.0
flake8==3.5.0
future==0.16.0
gunicorn==19.7.1
idna==2.7
mccabe==0.6.1
pew==1.1.5
Pillow==5.0.0
pipenv==9.0.3
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
requests==2.19.1
serializers==0.2.4
static3==0.7.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0